minor tidying
double quotes require the parser to look if there are variables to
interpolate. Using single quotes when there are none is a good
practice both for the visual hint to the developer and as a
parsing hint to the perl interpreter.

Also, sequential 'print' statements might require extra IO and
clutter the code. It is recommended to either concatenate the
strings (as this patch does) or, if the text is too big, using
heredocs.