get_kernel() and get_exploits() in, main() out
This is a minor maintainability update for the code. It provides
the get_kernel() and get_exploits() functions, and moves main()
to the top of the script. As such, developers are able to see
the complete logic of the code just by opening the file and
looking at the very first lines. The 'exit' at the end of the
code also guarantees that no extra data is processed, and the
rest are just auxiliary functions, created to provide extra
readability and maintainability.