making sure the output is properly sorted
hash key ordering should not be relied upon, as this may pose
a security issue. Depending on the hash size and machine, the
perl interpreter might optimize the hash variable by changing the
key order. In more recent perls (5.18+), the hash keys order is
randomized at every run to make programs more secure. This means
the order of exploits shown was not guaranteed to be the same
for every user, or even for the same user on different runs of
the script. This patch sorts the keys, forcing them to always
be displayed on the same order.